pyproject.toml
[tool.poetry]
name = "nfl_confidence"
version = "0.1.0"
description = ""
authors = ["Luke Ross <lukeross@umich.edu>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.3"
gspread = "^4.0.1"
numpy = "^1.21.2"
requests = "^2.26.0"
bs4 = "^0.0.1"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
loguru = "^0.7.2"
matplotlib = "^3.8.2"
tenacity = "^9.0.0"
pyyaml = "^6.0.2"
tqdm = "^4.66.5"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.10.0"
flake8 = "^6.1.0"
pytest = "^7.4.2"
coverage = "^7.3.2"
pytest-mock = "^3.12.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
